refactor: edit folder structure (#12)
