+++ src/main/java/com/example/waplan/goal/domain/FirstGoal.java
@@ -1,13 +1,17 @@
-package com.example.waplan.user.domain.goal;
+package com.example.waplan.goal.domain;
 
+import com.example.waplan.user.domain.User;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.util.List;
@@ -48,4 +52,9 @@ public class FirstGoal {
 
     @OneToMany(mappedBy = "firstGoal", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<SecondGoal> secondGoals;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
 }
